+++ aleo/artifacts/js/js2leo/wusdc_token.ts
@@ -9,7 +9,7 @@ import * as js2leo from "./common";
 export function gettokenLeo(token: token): tokenLeo {
   const result: tokenLeo = {
     owner: js2leo.privateField(js2leo.address(token.owner)),
-    amount: js2leo.privateField(js2leo.u64(token.amount)),
+    amount: js2leo.privateField(js2leo.u128(token.amount)),
     _nonce: js2leo.publicField(js2leo.group(token._nonce)),
   }
   return result;
